Move shared templates into their own package
