+++ package.json
@@ -56,7 +56,8 @@
         "./packages/catalog-server:build",
         "./packages/site-server:build",
         "./packages/site-content:build",
-        "./packages/site-client:build"
+        "./packages/site-client:build",
+        "./packages/site-templates:build"
       ]
     },
     "test": {
@@ -95,6 +96,7 @@
       "dependencies": [
         "./packages/site-client:check",
         "./packages/site-server:check",
+        "./packages/site-templates:check",
         "check:lint",
         "check:format"
       ]
