+++ packages/site-content/.gitignore
@@ -1,3 +1,2 @@
 /_site/
 /_dev/
-/templates/lib/
