+++ packages/site-content/package.json
@@ -9,19 +9,13 @@
   "scripts": {
     "build": "wireit",
     "build:prod": "wireit",
-    "build:dev": "wireit",
-    "build:templates": "wireit",
-    "build:templates:types": "wireit",
-    "check": "wireit",
-    "check:templates:types": "wireit"
+    "build:dev": "wireit"
   },
   "wireit": {
     "build": {
       "dependencies": [
         "build:prod",
-        "build:dev",
-        "build:templates",
-        "build:templates:types"
+        "build:dev"
       ]
     },
     "build:prod": {
@@ -34,7 +28,7 @@
         "_site"
       ],
       "dependencies": [
-        "build:templates",
+        "../site-templates:build:js",
         "../site-client:build:prod"
       ]
     },
@@ -49,46 +43,7 @@
         "_dev"
       ],
       "dependencies": [
-        "build:templates"
-      ]
-    },
-    "build:templates": {
-      "command": "find templates/src -name \"*.ts\" | xargs esbuild --color --outdir=templates/lib",
-      "dependencies": [
-        "../site-client:build"
-      ],
-      "files": [
-        "templates/src/**/*.ts"
-      ],
-      "output": [
-        "templates/lib/**/*.js"
-      ],
-      "clean": "if-file-deleted"
-    },
-    "build:templates:types": {
-      "command": "tsc --pretty",
-      "files": [
-        "templates/src/**/*.ts"
-      ],
-      "output": [
-        "templates/lib/**/*.d.ts",
-        "tsconfig.tsbuildinfo"
-      ],
-      "dependencies": [
-        "../catalog-api:build",
-        "../custom-elements-manifest-tools:build",
-        "../site-client:build"
-      ],
-      "clean": "if-file-deleted"
-    },
-    "check": {
-      "dependencies": [
-        "check:templates:types"
-      ]
-    },
-    "check:templates:types": {
-      "dependencies": [
-        "build:templates:types"
+        "../site-templates:build:js"
       ]
     }
   },
@@ -96,6 +51,7 @@
     "@11ty/eleventy": "^1.0.2",
     "@11ty/eleventy-navigation": "^0.3.5",
     "@webcomponents/internal-site-client": "^0.0.0",
-    "@webcomponents/internal-site-server": "^0.0.0"
+    "@webcomponents/internal-site-server": "^0.0.0",
+    "@webcomponents/internal-site-templates": "^0.0.0"
   }
 }
