+++ packages/site-server/package.json
@@ -40,7 +40,7 @@
       "dependencies": [
         "../site-client:build:dev",
         "../custom-elements-manifest-tools:build",
-        "../site-content:build:templates"
+        "../site-templates:build:js"
       ],
       "files": [
         "src/**/*.ts"
@@ -62,7 +62,7 @@
       ],
       "dependencies": [
         "../site-client:build:types",
-        "../site-content:build:templates:types"
+        "../site-templates:build:types"
       ]
     },
     "start:dev": {
@@ -117,6 +117,7 @@
     "@types/marked": "^4.0.8",
     "@web/dev-server": "^0.1.34",
     "@webcomponents/internal-site-content": "^0.0.0",
+    "@webcomponents/internal-site-templates": "^0.0.0",
     "google-auth-library": "^8.7.0",
     "koa": "^2.13.4",
     "koa-conditional-get": "^3.0.0",
