+++ packages/site-templates/src/base.ts
