+++ packages/site-templates/src/escape-html.ts
